fase-3: /db/* and /cache parity for C#, Java, Kotlin, GraalVM and Dart
Closes the REST contract across the remaining 15 implementations. All 37 now
answer /json, /db/simple, /db/complex and /cache with the shapes in
contracts/rest/canonical-payloads.md, and run the normative SQL behind /db/*.

graalvm/vertx was not running Vert.x. pom.xml names com.benchmark.vertx.Main
as main.class, and that class was a second, parallel server built on
com.sun.net.httpserver.HttpServer: /health answered a hardcoded "connected"
without touching Postgres or Redis, /db/simple returned an invented user,
/db/complex returned an empty data array, and /cache returned a fabricated
value. The real Vert.x server sitting next to it -- VertxServer, its router,
handlers and services -- was never started. Main is now a launcher for it.

Endpoints that reported the opposite of what happened:

  csharp (all 3)  "cached" was value.Contains("Cached value"), which is true
                  exactly when the value had just been generated
  graalvm/gspring "cached" was inferred from whether the stored string
                  contained today's date -- false for anything cached today

  Both now come from the cache service, which reports it directly.

Work that was in the measured path and should not have been:

  csharp (all 3)  await Task.Delay(50) on every /cache miss -- a hard ceiling
                  of ~20 req/s per connection
  graalvm/gmicronaut  three getOrSet calls per request, so three Redis round
                  trips, to infer one boolean
  kotlin/ktor     /db/simple and /db/complex concatenated their JSON by hand
                  into a StringBuilder, bypassing the serializer the rest of
                  the app -- and every other implementation -- was measured on
  graalvm/vertx   the real handler pulled 100 order rows and summed and
                  averaged them in Java; the database does it now

Response shapes: /db/simple returned {Id, Name, Email, CreatedAt, IsActive} in
C# -- names concatenated, no age, an IsActive computed from the row age; the
JVM implementations emitted snake_case keys; kotlin/ktor pinned snake_case
with @SerialName and carried an {order_count, avg_amount,
days_since_first_order} shape. /db/complex returned {orders, summary} or a
Java-side aggregate in several places.

SQL: java/spring and java/micronaut pasted `days` in with String.format;
kotlin/spring used string interpolation; dart/vaden and graalvm/vertx
concatenated it; six implementations joined order_items and aggregated
quantity*price, a materially heavier query than summing o.total_amount; none
had a tiebreak in ORDER BY, so rows with equal aggregates came back in
arbitrary order and no response was reproducible between runs.

Also: age is nullable in the schema but was read with GetInt32/getInt, which
throws or silently returns 0; kotlin/ktor hardcoded a 25-connection Hikari
pool instead of DB_POOL_MAX; and dart/vaden plus the Bun implementations read
a CACHE_TTL variable that is not in the ConfigMap.

Verification: dotnet build clean on all three C# projects; javac parse clean
on every changed Java file. No Maven, Gradle, kotlinc or Dart SDK on this
machine, so the JVM and Dart projects were not built whole -- that is Fase 6.

+++ src/csharp/FastEndpoints/Models/ComplexResult.cs
@@ -1,10 +1,12 @@
 namespace BenchmarkFastEndpoints.Models;
 
+// Mirrors UserOrderStats in contracts/grpc/benchmark.proto.
+// See contracts/rest/canonical-payloads.md.
 public class ComplexResult
 {
     public int UserId { get; set; }
     public string UserName { get; set; } = string.Empty;
-    public int TotalOrders { get; set; }
+    public long TotalOrders { get; set; }
     public decimal TotalValue { get; set; }
     public decimal AverageOrderValue { get; set; }
 }
